Add delete button
This commit also changed Article entity so that its parent entity is correctly mapped to Account entity. (To set an entity as a parent of another entity, the parent entity key must be saved as parent property of the child entity at child entity creation time. To properly configure a filter for an ancestor lookup, the filter ancestor property must be set at filter creation time.)

This change was necessary because without proper parent-child relationship, the query operation returned deleted articles because the query operation was not structured for strong consistency.